Restore support for round tripping leap seconds
This is a bit of hack until `time` has built-in support for leap
seconds. When we detect a leap second in `GeneralizedTime`, we set the
`is_leap_second` flag to true. When serializing to bytes, we check the
flag and bump the seconds to 60 if it's set. A debug assert ensures
we'll notice if the flag is set incorrectly.