Switch from `chrono` to `time` (#61)
* Switch from `chrono` to `time`

There is unsoundness in `chrono`
(https://rustsec.org/advisories/RUSTSEC-2020-0159), and although we
don't use the affected APIs, the advisory still fires. `time` ≥3 does
not have this unsoundness, and supports almost‡ all the necessary APIs,
so we can quite simply swap it out.

Since the `chrono` types are used in some APIs, those signatures have
had to change, making this a breaking change. Since it's already a
breaking change, the feature is also renamed from `chrono` to `time`.

‡ `time` does not support leap seconds in their representation or
parsing. In this commit, we simply ignore leap seconds, which also
constitutes a breaking change since values with leap seconds will parse
differently (they parse to the previous second).

* Exclude optional features from MSRV

Some optional dependencies, for example `time`, have a more strict MSRV
than this crate. Rather than having to upgrade our MSRV or pin outdated
versions of dependencies, we can instead explicitly exclude optional
features from the MSRV.

The test CI workflow has been updated to not run tests with features on
MSRV.

* Update datetime references in docs to `OffsetDateTime`

This is a better match for the `time` terminology.

* Restore support for round tripping leap seconds

This is a bit of hack until `time` has built-in support for leap
seconds. When we detect a leap second in `GeneralizedTime`, we set the
`is_leap_second` flag to true. When serializing to bytes, we check the
flag and bump the seconds to 60 if it's set. A debug assert ensures
we'll notice if the flag is set incorrectly.